Fix calibration structure alignment and remove warning about anonymous structs.

+++ persist_config.h
@@ -23,8 +23,8 @@ extern bool        persistent_get_cal_type(unsigned adc, cal_type_t * type);
 extern bool        persistent_get_cal(unsigned adc, float * scale, float * offset, const char ** unit);
 extern bool        persistent_set_cal(unsigned adc, float scale, float offset, const char * unit);
 
-extern bool        persistent_get_exp_cal(unsigned adc, float ** cals, unsigned * count, const char ** unit);
-extern bool        persistent_set_exp_cal(unsigned adc, float * cals, unsigned count, const char * unit);
+extern bool        persistent_get_exp_cal(unsigned adc, double ** cals, unsigned * count, const char ** unit);
+extern bool        persistent_set_exp_cal(unsigned adc, double * cals, unsigned count, const char * unit);
 
 extern bool        persistent_set_use_cal(bool enable);
 extern bool        persistent_get_use_cal();
